Move `Tapioca::Compilers::Gem::Events` to `Tapioca::Gem::Events`
