+++ lib/tapioca/internal.rb
@@ -17,7 +17,7 @@ require "tapioca/gemfile"
 require "tapioca/executor"
 require "tapioca/symbol_table_parser"
 require "tapioca/symbol_loader"
-require "tapioca/compilers/gem/events"
+require "tapioca/gem/events"
 require "tapioca/compilers/gem/listeners"
 require "tapioca/compilers/requires_compiler"
 require "tapioca/compilers/symbol_table_compiler"
